refactor: AggregatorをGtAggregatorとCrossAggregatorに分離
責務が過大だったAggregatorクラスをGT集計専用のGtAggregatorと
クロス集計1軸分を担当するCrossAggregatorに分割。
aggregate()関数でGT→各クロス軸をPromise.allで並列実行しマージする構成に変更。
